fix: :bug: O seeder da tabela de produtos foi modificado para consertar um valor errado

+++ back-end/src/database/models/product.js
@@ -7,7 +7,7 @@ const Product = (sequelize, DataTypes) => {
         allowNull: false,
       } ,
       name: DataTypes.STRING,
-      price: DataTypes.FLOAT,
+      price: DataTypes.DECIMAL,
       url_image: DataTypes.STRING,
     }, {
       tableName: 'Products',
